Initial commit
This initial commit includes the first draft of the remake, which
consists of a player controller that triggers a moveable script
that causes the player game object to move one tile in a certain
direction.

The moveable script also works with other moveable scripts in order
to achieve chained moveables, which gives the effect that the player
is pushing tiles, and also gives them the ability to push multiple
tiles.

This draft version includes some smoke-and-mirrors that immitates
the functionality of Baba is You where-by moving word tiles will
affect the game logic.

This version also has sprites that use a top-left pivot point. This
was done to make it easier to slot our sprites into a square grid,
but resulted in complications when trying to use other editor features
such as tilemaps and rule-tiles.